Gent-Davenport-Maintenance/func-api/requirements.txt
azure-functions
azure-identity
azure-ai-projects==2.0.0b1
openai
